Bug 483292 - [Benchmark] long must be used to store memory usage

+++ org.eclipse.modisco.infra.discovery.benchmark.core/src/org/eclipse/modisco/infra/discovery/benchmark/core/internal/impl/EventAndMemoryRecorder.java
@@ -1,11 +1,13 @@
 /*******************************************************************************
- * Copyright (c) 2012 INRIA. All rights reserved. This program and the
- * accompanying materials are made available under the terms of the Eclipse
- * Public License v1.0 which accompanies this distribution, and is available at
+ * Copyright (c) 2012, 2015 INRIA and Mia-Software
+ * All rights reserved. This program and the accompanying materials are made 
+ * available under the terms of the Eclipse Public License v1.0 which 
+ * accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  * 
- * Contributors: Guillaume Doux - INRIA - Initial API and implementation
- * 
+ * Contributors:
+ *     Guillaume Doux (INRIA) - Initial API and implementation
+ *     Grégoire Dupé (Mia-Software) - Bug 483292 - [Benchmark] long must be used to store memory usage
  ******************************************************************************/
 
 package org.eclipse.modisco.infra.discovery.benchmark.core.internal.impl;
@@ -29,7 +31,7 @@ public class EventAndMemoryRecorder implements IEventListener {
 	
 	private long startTime;
 	private long stopTime;
-	private int maxMemoryUsed;
+	private long maxMemoryUsed;
 	private boolean measureMemoryUse;
 	
 	private List<MemoryMeasurement> memoryMeasurements;
@@ -100,7 +102,7 @@ public class EventAndMemoryRecorder implements IEventListener {
 	 * Max memory used getter 
 	 * @return the max memory used
 	 */
-	public int getMaxMemoryUsed() {
+	public long getMaxMemoryUsed() {
 		return this.maxMemoryUsed;
 	}
