+++ org.eclipse.modisco.infra.discovery.benchmark.metamodel/src/org/eclipse/modisco/infra/discovery/benchmark/metamodel/internal/benchmark/MemoryMeasurement.java
@@ -1,14 +1,13 @@
 /**
- * *******************************************************************************
- *  * Copyright (c) 2012 INRIA. All rights reserved. This program and the
- *  * accompanying materials are made available under the terms of the Eclipse
- *  * Public License v1.0 which accompanies this distribution, and is available at
- *  * http://www.eclipse.org/legal/epl-v10.html
- *  * 
- *  * Contributors: Guillaume Doux - INRIA - Initial API and implementation
- *  * 
- *  ******************************************************************************
+ * Copyright (c) 2012, 2015 INRIA and Mia-Software.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
  * 
+ * Contributors:
+ *      Guillaume Doux (INRIA) - Initial API and implementation
+ *      Grégoire Dupé (Mia-Software) - Bug 483292 - [Benchmark] long must be used to store memory usage
  */
 package org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchmark;
 
@@ -20,10 +19,10 @@ package org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchma
  *
  * <p>
  * The following features are supported:
+ * </p>
  * <ul>
  *   <li>{@link org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchmark.MemoryMeasurement#getMemoryUsed <em>Memory Used</em>}</li>
  * </ul>
- * </p>
  *
  * @see org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchmark.BenchmarkPackage#getMemoryMeasurement()
  * @model
@@ -46,12 +45,12 @@ public interface MemoryMeasurement extends Event {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Memory Used</em>' attribute.
-	 * @see #setMemoryUsed(int)
+	 * @see #setMemoryUsed(long)
 	 * @see org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchmark.BenchmarkPackage#getMemoryMeasurement_MemoryUsed()
 	 * @model
 	 * @generated
 	 */
-	int getMemoryUsed();
+	long getMemoryUsed();
 
 	/**
 	 * Sets the value of the '{@link org.eclipse.modisco.infra.discovery.benchmark.metamodel.internal.benchmark.MemoryMeasurement#getMemoryUsed <em>Memory Used</em>}' attribute.
@@ -61,6 +60,6 @@ public interface MemoryMeasurement extends Event {
 	 * @see #getMemoryUsed()
 	 * @generated
 	 */
-	void setMemoryUsed(int value);
+	void setMemoryUsed(long value);
 
 } // MemoryMeasurement
